backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.1
opencv-python==4.8.1.78
transformers==4.34.1
torch==2.1.0
pytesseract>=0.3.8
python-jose==3.3.0
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
starlette>=0.20.0
passlib==1.7.4